split up view routes
