+++ pephub/main.py
@@ -17,6 +17,10 @@ from .routers.api.v1.user import user as api_user
 from .routers.api.v1.search import search as api_search
 from .routers.auth.base import auth as auth_router
 from .routers.views.base import views as views_base
+from .routers.views.search import search as views_search
+from .routers.views.project import project as views_project
+from .routers.views.namespace import namespace as views_namespace
+from .routers.views.user import user as views_user
 from .routers.eido import router as eido_router
 from .const import STATICS_PATH, EIDO_PATH
 
@@ -72,6 +76,10 @@ app.include_router(api_project)
 app.include_router(api_search)
 app.include_router(auth_router)
 app.include_router(views_base)
+app.include_router(views_user)
+app.include_router(views_search)
+app.include_router(views_project)
+app.include_router(views_namespace)
 app.include_router(eido_router)
 
 # mount the landing html/assets
